Modified the BasicBufferMgr to support efficient empty frame finding.
Utilize ArrayList of unpinned (empty) frames (buffers) to which Buffers are added if they are unpinned/empty and removed from when they are pinned. When asking for an empty frame, the first element in this ArrayList is return if the list is non-empty. Tested SimpleDB startup as well as sample scripts and test db warm-up program and all worked with modifications.